backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
blinker==1.7.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==5.2.2
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.1
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.10
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==2.1.5
pillow==11.3.0
PyJWT==2.10.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
sql==2022.4.0
SQLAlchemy==2.0.27
sqlparse==0.5.3
typing_extensions==4.10.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.0.1
python-dotenv==1.0.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
gunicorn
channels==4.0.0
channels-redis==4.1.0



